Add completionTime to cluster status
Used to fill in `stop_time` in the cluster models returned by the api
server.
