+++ dask-gateway-server/dask_gateway_server/backends/kubernetes/utils.py
@@ -1,5 +1,7 @@
 import asyncio
 import json
+import sys
+from datetime import datetime, timezone
 
 from traitlets import Dict, Unicode, Any, Int
 from traitlets.config import LoggingConfigurable
@@ -12,6 +14,25 @@ from ...utils import cancel_task
 
 # Monkeypatch kubernetes_asyncio to cleanup resources better
 client.rest.RESTClientObject.__del__ = lambda self: None
+
+
+if sys.version_info[:2] >= (3, 7):
+
+    def parse_k8s_timestamp(ts):
+        t = datetime.fromisoformat(ts[:-1])
+        return int(t.replace(tzinfo=timezone.utc).timestamp() * 1000)
+
+
+else:
+
+    def parse_k8s_timestamp(ts):
+        t = datetime.strptime(ts, "%Y-%m-%dT%H:%M:%SZ")
+        return int(t.replace(tzinfo=timezone.utc).timestamp() * 1000)
+
+
+def k8s_timestamp():
+    t = datetime.now(timezone.utc)
+    return t.strftime("%Y-%m-%dT%H:%M:%SZ")
 
 
 async def watch(method, *args, **kwargs):
